requirements.txt
pysqlite3==0.4.2
Flask-SQLAlchemy==2.4.1
Flask==1.1.1
